test(range): migrate basic
