+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-default-ios-ltr-Mobile-Chrome-linux.png
