+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-default-md-ltr-Mobile-Safari-linux.png
