+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-diff-ios-ltr-Mobile-Safari-linux.png
