+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-dual-knobs-ios-ltr-Mobile-Firefox-linux.png
