+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-dual-knobs-md-ltr-Mobile-Chrome-linux.png
