+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-pin-ios-ltr-Mobile-Firefox-linux.png
