+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-pin-md-ltr-Mobile-Safari-linux.png
