+++ core/src/components/range/test/basic/range.e2e.ts-snapshots/range-ticks-md-ltr-Mobile-Firefox-linux.png
